+++ frontend/src/components/molecules/Activity.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+function Activity(shouldMin, shouldMax, lastdata) {
+  return (
+    <div>
+        
+
+    </div>
+  )
+}
+
+export default Activity