+++ frontend/vite.config.js
@@ -3,5 +3,19 @@ import react from '@vitejs/plugin-react'
 
 // https://vite.dev/config/
 export default defineConfig({
-  plugins: [react()],
+  plugins: [
+    react({
+      include: '**/*.{jsx,js}',
+    }),
+  ],
+  base: '/static/',
+  build: {
+    outDir: 'dist',
+    assetsDir: '',
+  },
+  server: {
+    proxy: {
+      '/api': 'http://localhost:8000',
+    },
+  },
 })